% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterModule.R
\name{filterModule}
\alias{filterModule}
\alias{filterModuleOuput}
\title{A shiny module for automatically generating filter widgets for a table.}
\usage{
filterModule(input, output, session, data)

filterModuleOuput(id)
}
\description{
A module to be called in the server portion of a shiny app.

Filter output for a shiny app.
}
\section{Functions}{
\itemize{
\item \code{filterModuleOuput}: The output for the ui--call where you want the filter widgets.
}}
\examples{
# main ui
ui <- fixedPage(

 filterModuleOuput("id1"),

 DT::dataTableOutput("table")
)

# main server
server <- function(input, output, session) {
 data <- iris
 filter_data <- callModule(filterModule, "id1", data = data)

 output$table <- DT::renderDataTable({
   filter_data()
 })
}

# run app
shinyApp(ui, server)
}

